WIP: Support DOMAction-level routing
We have a distinct use case where we want to route actions across
a system. In this scenario, a next hop provider registers itself as the
local provider with the appropriate cost.

This capability is also exposed as the simplest way of registering
binding-level implementations -- where we want to register to all
actions on the operational datastore.

Unfortunately our current interface does not allow that, as we strictly
require at least DOMActionInstance. This is a modeling accident, as we
need a datastore type -- which we can easily provide.

Update DOMActionProviderService with new methods to allow for this use
case and add proper tests for action routing.

// FIXME: finish implementation

JIRA: MDSAL-681
Change-Id: I6daf70b28347fab88c8dc467fce267d8ba71e57d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>