Move createBindingArg() logic to BindingRuntimeContext
There is an intricate situation here, for which binding-runtime-api
is a better place. Rather than performing piece-meal resolution, get
what we need from BindingRuntimeContext in one go, adjusting callers
accordingly.

With this patch, it is up to binding-runtime-api to coordinate with
mdsal-binding-generator the information required to remove use of
isAddedByUses().

It seems we will be carrying some information between parent and child,
esp. regarding uses and augments, hence we introduce an explicit class
in the hope it will serve as a common meeting place. This typed
indirection is better that using Map.Entry for a complex return.

JIRA: MDSAL-697
Change-Id: Iae12f1749e7b4b76abc6bfc222c7ca7d2f5769cb
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>