WIP: Add nonnullFoo() getters for non-semantic objects
Add non-null getters for containers without presence
statement which returns lazy-initiated containers'
empty instances.

For this purpose we need to create containers' empty
instances in corresponding builder classes.

In builders empty instance is hidden behind
a static method and is only initialized when
empty() method is invoked for the first time.

JIRA: MDSAL-673
Change-Id: I3a5a346b32b89e7c28cad31536593824dee3c232
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>