Enforce explicit generator linkage
We are keeping the original pointer as a simple nullable field. This can
result in us confusing an unresolved generator with an original, leading
to potential badness.

Refactor the AbstractExplicitGenerator to be a tri-state field, which
flushes out a deficiency in our resolution logic, tripping up one of our
unit tests.

Change-Id: Ie67fa4d08d0887f301948e3d03d846ed9ee1d628
JIRA: MDSAL-718
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>