Inject rpc input QName into LazySerializedContainerNode
When ExecuteRpc message is sent with LazySerializedContainerNode
as 'input' parameter, during serialization is invoked ContainerNode's
getIdentifier() method that returns NodeIdentfier of rpc,
not input. So OpsInvoker receives the message with incorrect 'input'
parameter and downstream we get error due to it.

We need to inject rpc input Qname into LazySerializedContainerNode
to get NodeIdentifier of rpc input from getIdentifier() method.

JIRA: INTTEST-130
Change-Id: I77c870fe91a758037bd27e54607e778255d611fc
Signed-off-by: Dominik Vrbovsky <dominik.vrbovsky@pantheon.tech>