Map bits type to primitive boolean
The main change is in the AbstractTypeObjectGenerator.java,
returned type for bits was set to primitive boolean.
Generation of the getDefaultInstance() method in the ClassTemplate.xtend
had to be adjusted as well, as the null value obviously
does not make sense in the case of primitive type.
Related test classes had to be changed as well

JIRA: MDSAL-744
Change-Id: Ic6423f67679475fced73ea68ed4f447b241230f7
Signed-off-by: ivan.martiniak <ivan.martiniak@pantheon.tech>