Add BitsTypeObject
The new interface BitsTypeObject, besides grouping bits type objects,
helps clarify which bits are restricted.

Method validValues() returns constant containing valid bits.
The logic of the creation of this constant is contained in the
AbstractTypeObjectGenerator.java.

Method getValue() was included in this interface and its generation
was adjusted for restricted and standard bits types separately.

JIRA: MDSAL-743
Change-Id: I74820e7a041e137902c774655517ea8d1e4bccd9
Signed-off-by: Ivan Martiniak <ivan.martiniak@pantheon.tech>