Add support for registering single-RPC services
We are currently providing an aggregated RpcService generated class,
which is binding the invocation model to that class -- without the
ability to provide an implementation of a single RPC. It also
complicates the invocation path from DOM, since we need to perform
two lookups: one in DOM, where we route the invocation towards the
RpcServiceAdapter and then a method-based lookup to dispatch it in
the proxy.

Furthermore this prevents the possibility of adapting the invocation
contract, as the codegen is always involved.

We are now generating interfaces representing individual RPCs, just like
we do for actions. This provides a more granular access and notably
allows mdsal-binding-api to provide invocation contract adapters -- thus
disconnecting it from codegen.

JIRA: MDSAL-777
Change-Id: I114bd3795ba353da57e7f96d5db9680986ecb5f6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>