Simplify YangModuleInfoScanner
We will be tracking YangFeatureProviders, so
List<ObjectRegistration<YangModuleInfo>> is not quite accurate. Track
only a simple Registration and make sure we wrap the return from
ModuleInfoSnapshotResolver.

Also improve the tracking logic a bit: do not track bundles which do not
contain YangModuleInfos, but make a special exception for the system
bundle.

JIRA: MDSAL-765
Change-Id: I245f25c72152f7e49817c316f63ace557b4c56c0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>