Refactor RpcRoutingStrategy
RpcRoutingStrategy is not really a nice construct -- it essentially
extracts context information, not much more. As such, dealing with two
subclasses is a drag.

Intruduce ContentRoutedRpcContext to capture the interesting information
in a record and adjust users to deal with the fact that normal RPCs do
not have this construct.

Change-Id: I2c8d24a61e6905063c9eb93fb0513dffe4831b66
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>