Update CurrentAdapterSerializer.getRpcMethodToSchema()
We are moving less reliance on BindingReflections and RpcDefinition.
Refactor getRpcMethodToSchema() to use RpcRuntimeTypes instead, from
whence we can acquire better information -- for example
RpcEffectiveStatement.

Also update LazySerializedContainerNodeTest to not rely on this method,
as it can get its inputs in a much smoother manner.

This is another step to eliminating BindingReflections from invocation
strategy creation, which we will need, as single RPCs just do not have
the type capture we need for the input class.

JIRA: MDSAL-777
Change-Id: I7382e2a01ee8ee0d0e611286f9dcef5d92e6acdf
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>