Change action input/output naming strategy
Users may want to implement multiple actions in a single source file, to
which extent the placement and bare names of Input and Output are rather
bad -- they cannot be imported more than once, leading to clutter.

RPC input/output do not suffer from this, as they combine their names
with the name of the RPC -- leading to to classes which can easily be
imported.

Change-Id: I59e9926d7c5504519e9c9ba43c6915f6e5808ad6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>