Eliminate RpcMethodInvoker
RpcMethodInvoker is a useless indirection around a MethodHandle with
a known signature. Eliminate this indirection and cache all
MethodHandles for a particular Rpc, passing each of them the its
individual LegacyDOMRpcImplementationAdapter.

This allows LegacyDOMRpcImplementationAdapter to fuse the delegate
object into its internal MethodHandle -- hence we do not need to track
the delegate separately.

JIRA: MDSAL-86
Change-Id: I9b89d353fc61e313c82e997ffeddbbb473bbd07f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>