Disconnect AbstractExplicitGenerator from CopyableNode
The two bits we are exposing through this interface are important,
but they should be compuated separately from the effective model.

Expend two booleans to track them eagerly and disconnect the access
methods from CopyableNode. At the end of the day we should be receiving
this information as Provenance enum.

Change-Id: Ic1c84c2432afcaedde417a1ef89fe6cc207742a2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>