Fix typedef/bits code generation
bits/typedef seemed to be generating wrong code. Added test to check if the generated code contains desired method. Needed to add new method to FileSearchUtil that only checks if the generated file contains separate lines, not the whole method as the joined string of lines had quotation marks prefixed with a backslash (for unknown reason), which made the tests not pass.

JIRA: MDSAL-807
Change-Id: I4618269fff1fc284b65e9efb1d49501248d233c3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
Signed-off-by: matus.matok <matus.matok@pantheon.tech>