Generated unions patterns enforcement
Changes in .yang files were done because revision is not mandatory statemet, so it may not be present. This regex allows it to math for both present and non-present.
Adding ^$ to the pattern was not successful, since the way enforcment string is generated adds extra '\' to those symbols which doesnt allow it to match.

JIRA: MDSAL-405
Change-Id: I63880fba2aac98d04b42b208d0b74e1fd49b8f4e
Signed-off-by: Oleksandr Panasiuk <oleksandr.panasiuk@pantheon.tech>