WIP: Make InstanceIdentifierBuilder variadic to KeyedInstanceIdentifier
Through analysis, we figured out that for example this call:
InstanceIdentifier.builder(List1.class).child(List12.class, new List12Key(12)).build();
would result in InstanceIdentifier, instead of KeyedInstanceIdentifier because the .builder() (or any other .child() method in the chain could do the same) caused the builder to change its internal state to 'wildcarded = true;' as List1 is Identifiable<?>. And if the state is wildcarded, we can only yield an InstanceIdentifier on the .build() method, despite being in KeyedInstanceIdentifierBuilder as per task goal. To patch up the leak in logic we did the following:

1) added new classes WildcardedInstanceIdentifier and WildcardedInstanceIdentifierBuilder to surrogate the function of the field wilcarded in InstanceIdentifier and its builder. Hence, if the state of a builder was to be set to wilcarded, we now return WildcardedInstanceIdentifierBuilder which only yields WilcardedInstanceIdentifier on its .build() method.
2) added new class KeyedInstanceIdentifierBuilder as in the original task goal.
3) added new abstract class AbstractInstanceIdentifierBuilder to hold some common functionality of the builders.
4) added tests to define the desired behavior of builders.
5) added interfaces for WildcardedInstanceIdentifierBuilder and KeyedInstanceIdentifierBuilder similiar to InstanceIdentifierBuilder. Renamed new classes WildcardedInstanceIdentifierBuilder to WildcardedInstanceIdentifierBuilderImpl (similarly for KeyedInstanceIdentifierBuilder) and made them implement their respective interfaces.
6) considered deprecation of InstanceIdentifier's .child() methods which do exactly the same thing as .child() methods of builders, without the necessity to use a builder, which implies we would need to do the same changes twice for no added benefit. However decided not to do so, for runtime complexity reasons.

If this prototype were to be finished, the following would need to be done:

//TODO:

7) go through the code base and exchange .child() for .wildcardChild() methods where necessary.
8) write proper tests as the patch is quite extensive, so we might have missed many edge cases

It was decided to leave this patch as it is. Might serve as an inspiration for the future.

JIRA: MDSAL-798
Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Change-Id: I404df6aa1f3c79f259ea705adedd58a9d466cb78