Improve DOMMountPointServiceImpl error message
The checkState() we perform is rather ugly, as it does not provide
enough diagnostic information. We already have a nicer checkState()
in the register() path. Extract it to a common method and reuse it from
both places.

Also improve safety by checking for null path before anything and also
take the lock to ensure serialization with others.

JIRA: NETCONF-621
Change-Id: I55845f9bbd7b3f4dbfa06a91724abcb601dc4f7f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>