Clean up BindingToNormalizedStreamWriter
Clarify field name: the rootNodeSchema is really a
DataContainerCodecContext the context corresponding to the root object,
hence it should be called rootSchema.

We have multiple calls to current(), which end talking to the Deque,
which is not necessary. We can just access it once, which makes type
checking more obvious.

Finally ditch checkArgument() calls and prefer instanceof patterns,
which makes the type checks more obvious (to the compiler at least).

Change-Id: I472f0f3eec1f83c1755c91c6cbd0b24b77ae9d9a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>