Change YangDataNamingStrategy to not confuse the JVM
Problem here was that the JVM uses $ in the end of the name to identify Nested/Inner classes, which in our case caused it to look for CyrillicNaming.class inside привет class but not in the привет$ dir.
This issue did not appear for all the team members probably because we have different Linux distributives/versions and JVM asks for a system file splitter to properly identify the filepath.
It seems that there is something different in Ubuntu 22.04 which causes JVM to fail finding the appropriate filepath to class if it is located in the directory with the dollar sign in the end.

JIRA: MDSAL-814
Signed-off-by: Oleksandr Panasiuk <oleksandr.panasiuk@pantheon.tech>
Change-Id: I1c328839edca39a46d5d8ea5ec5414d1e38daa99