Clean up IETF module names
The primary key is the IETF module name, make sure that is reflected in
packaging. The basic idea here is that differing revisions and up
reflecting diffent package names.

This patch allows for differing revisions to exist based on standard
names and the revision -- whereas previously we used the document as a
starting point for the namespace.

JIRA: MDSAL-752
Change-Id: I3ea3f910d2c07dad8396c54fa03044ac6a4ad62b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>