Add leafref cycle detection
Currently, we only detect self-referencing leafrefs,
but there is no detection for leafref cycle chains.

Add detection for circular leafref chains.
This is achieved by tracking the furthest generator in leafref
chain referenced by given leafref at the point when
resolveLeafref() is called.
At point when leafref is being resolved, it is compared to
this furthest generator in leafref chain its target is
part of. And if no cycle is detected current leafref is
updated as resolved and part of the chain with given
the furthest generator in chain.

If the target of leafref which is currently being resolved
is not yet resolved or is not leafref it is set as the furthest
in chain.

JIRA: MDSAL-505
Change-Id: Id8dcc03db16a373ca3ec85bc91900da81bace36f
Signed-off-by: Samuel Schneider <samuel.schneider@pantheon.tech>