Changed $YangModuleInfoImpl and $YangModelBindingProvider generation package and name
The naming of these two classes is rather unfortunate, as it starts with $ due to the requirement not to conflict with other generated classes. So they were renemed to avoid dollar sign usage.
Also the root package for theese classes was changed from org.opendaylight.yang.gen.v1 to org.opendaylight.yang.svc.v1.
A lot of test classes are touched in this patch but no logic were changed there, its just providing a new path for YangModuleInfoImpl.

JIRA: MDSAL-836
Change-Id: If9e43d21063c2fc48d8adbf0ccfc2406fe760054
Signed-off-by: Oleksandr Panasiuk <oleksandr.panasiuk@pantheon.tech>