WIP: Change instance-identifier mapping
InstanceIdentifier does not cover the entire value space of YANG's
instance-identifier value space, which is a problem, as we just cannot
emit proper accessors for YANG-modeled data when the instance identifier
points to a particular item.

This patch adds an alternative, BindingInstanceIdentifier, which models
the full equivalent of YangInstanceIdentifier.

// FIXME: add builders
// FIXME: deprecate InstanceIdentifier et al.

JIRA: MDSAL-840
Change-Id: I9046128a4057518d09337d5d69289e86ab389dce
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>