Clean up mdsal-singleton-dom-impl
Fold the multitude of implementation classes into a single
EOSClusterSingletonServiceProvider.

Furthermore do not expose close() from the implemented interface, so
that users do not end up shutting the implementation down.

Eliminate ServiceLoader integration, as we are a stateful component
and therefore want to explicit lifecycle.

The shutdown process is made synchronous, so that close() waits for
everything to settle down.

JIRA: MDSAL-843
Change-Id: I0d137d03d184de07e21e5a9a4329855a021dd433
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>